% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsr-package.R
\docType{package}
\name{bmsr-package}
\alias{bmsr-package}
\alias{bmsr}
\title{The 'bmsr' package.}
\description{
The bmsr package implements joint regression from multiple data sources in a Bayesian framekwork. The package provides implementation for btoh single-task and multi-task regression. The model is implemented using STAN and interface is provided using R programming language. Options for training the model using both NUTS sampler and variational inference are provided. The package is structured for ease of use and the included demo shows the model execution on real-life as well as simulated datasets.
}
\references{
Brian S. White, Suleiman A. Khan, Mike J Mason, Muhammad Ammad-ud-din, Swapnil Potdar, Disha Malani, Heikki Kuusanmäki, Brian J. Druker, Caroline A Heckman, Olli Kallioniemi, Stephen E Kurtz, Kimmo Porkka, Cristina E. Tognon, Jeffrey W. Tyner, Tero Aittokallio, Krister Wennerberg, Justin Guinney,
\emph{Bayesian multi-source regression and monocyte-associated gene expression predict BCL-2 inhibitor resistance in acute myeloid leukemia},
<To Appear>, (2021)
}
