% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_utils.R
\name{demo_bmsr}
\alias{demo_bmsr}
\title{demo multi-source model training and interpretative plots}
\usage{
demo_bmsr()
}
\description{
\code{demo_bmsr} trains bmsr on random multi-source regression dataset.
}
\examples{

#for multi-source regression
demo_bmsr()

}
