% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_utils.R
\name{demo_bmsr_lapatinib_gdsc_ccle}
\alias{demo_bmsr_lapatinib_gdsc_ccle}
\title{demo real data training and result visualization}
\usage{
demo_bmsr_lapatinib_gdsc_ccle()
}
\description{
\code{demo_bmsr_lapatinib_gdsc_ccle} trains bmsr model on real dataset from GDSC and CCLE datasets from the PharmacoGx package.
The demo trains a model for lapatinib response predictions.
}
\examples{

#model lapatanib response from GDSC and CCLE
demo_bmsr_lapatinib_gdsc_ccle()

}
