% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_utils.R
\name{demo_bmsr_venetoclax_tavor_ohsu}
\alias{demo_bmsr_venetoclax_tavor_ohsu}
\title{Demo applying BMSR for integrative analysis of Tavor and Beat AML/OHSU datasets to discover biomarkers of venetoclax response.}
\usage{
demo_bmsr_venetoclax_tavor_ohsu()
}
\description{
\code{demo_bmsr_venetoclax_tavor_ohsu} trains bmsr model on real dataset from Tavor and Beat AML/OHSU datasets downloaded from
ORCESTRA and harmonized with PharmacoGx. The demo trains a model to predict biomarkers of venetoclax response in AML.
}
\examples{

#modelv venetoclax response from Tavor and Beat AML/OHSU datasets
demo_bmsr_venetoclax_tavor_ohsu()

}
