% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_exceedance.R
\name{dirichlet_exceedance}
\alias{dirichlet_exceedance}
\title{Compute exceedance probabilities}
\usage{
dirichlet_exceedance(alpha, Nsamp = 1e+06)
}
\arguments{
\item{alpha}{Dirichlet parameters (usually the posterior estimate).}

\item{Nsamp}{number of samples drawn.}
}
\value{
Exceedance probabilities.
}
\description{
This function compute exceedance probabilities by simulating the multivariate Dirichlet observations using the same approach implemented in SPM 12.
}
