% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utils.R
\name{download.psets}
\alias{download.psets}
\title{Download PharmacoGx-compatible PSets (i.e., URL-linked RDS files)}
\usage{
download.psets(pset.links)
}
\arguments{
\item{pset.links}{a named list of URLs to PSets, i.e., list(dataset1 = url1, dataset2 = url2, etc.)}
}
\description{
\code{download.psets} trains bmsmtr on random multi-source multi-task regression dataset.
}
