% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FE_null.R
\name{FE_null}
\alias{FE_null}
\title{Calculate free energy for H0 (equal model frequencies in the population)}
\usage{
FE_null(m)
}
\arguments{
\item{m}{N by K (N subjects by K models) matrix of log-model evidence.}
}
\value{
Free energy.
}
\description{
THis function derives the free energy of the 'null' (H0: equal model frequencies). This routine has been copied from the VBA_groupBMC function of the VBA toolbox http://code.google.com/p/mbb-vb-toolbox/ which was written by Lionel Rigoux and J. Daunizeau. See Equation A.17 in Rigoux et al.
}
