% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_utils.R
\name{generateSyntheticData}
\alias{generateSyntheticData}
\title{generate multi-source toy data for regression}
\usage{
generateSyntheticData(S = 2, nY = c(40, 80), dX = 50, dY = 1, ft = 10)
}
\arguments{
\item{S}{is a scaler representing the desired number of sources.}

\item{nY}{is a vector representing the desired number of samples in each source.}

\item{dX}{is a scaler representing the desired number of dimensions in X (inputs).}

\item{dY}{is a scaler representing the desired number of dimensions in Y (outputs). dY > 1 refers to multi-task datasets.}

\item{ft}{is a scaler representing the desired number of active features in X.}
}
\value{
A list containing the following elements:
\item{data} which is a list containing the Y and X data matrices
\item{Beta} the beta parameters used to generate the data
}
\description{
\code{generateSyntheticData} simualtes random multi-source multi-task regression dataset.
}
