% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsr.R
\name{getBeta.bmsr.stan}
\alias{getBeta.bmsr.stan}
\title{get beta posterior of the bmsr model}
\usage{
getBeta.bmsr.stan(out)
}
\arguments{
\item{out}{is trained STAN model.}
}
\value{
beta matrix containing source specific beta parameters.
}
\description{
\code{getBeta.bmsr.stan} extracts the posterior values of source specific beta parameters.
}
