% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsr.R
\name{getPosterior}
\alias{getPosterior}
\title{baseline function to get posterior}
\usage{
getPosterior(file = NULL, out)
}
\arguments{
\item{file}{is the stan file name containig the stan code.}

\item{out}{is trained STAN model.}
}
\value{
post is a list containing posterior of all model weights.
}
\description{
\code{getPosterior} extracts the posterior values from mode output.
}
