% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpingFunctions.R
\name{meanNorm}
\alias{meanNorm}
\title{mean normalize a data matrix}
\usage{
meanNorm(mat, FFSC, verbose = FALSE)
}
\arguments{
\item{mat}{matrix to mean normalize}

\item{FFSC}{the index of first column to standardize. All columns including this and after are normalized.}

\item{verbose}{print logs or not. Defaults to FALSE.}
}
\value{
a list containing the following:
\item{mat} column-wise mean normalized matrix.
\item{cm} vector of column means used for mean normalization.
}
\description{
\code{meanNorm} performs column-wise mean=0 normalization of a matrix for all columns starting from \code{FFSC}
}
\keyword{internal}
