% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsr.R
\name{posterior.bmsmtr.stan}
\alias{posterior.bmsmtr.stan}
\title{get posterior of the bmsmtr model weights}
\usage{
posterior.bmsmtr.stan(out)
}
\arguments{
\item{out}{is trained STAN model.}
}
\value{
post is a list containing following posterior weights.
\item{beta}{matrix containing source specific beta parameters,}
\item{W}{matrix of multi-task parameters,}
\item{tau}{global scaling factor learned,}
\item{sigma}{noise parameter.}
}
\description{
\code{getPosterior} extracts the posterior values from mode output.
}
