% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsr.R
\name{posterior.bmsr.stan}
\alias{posterior.bmsr.stan}
\title{get posterior of the bmsr model weights}
\usage{
posterior.bmsr.stan(out)
}
\arguments{
\item{out}{is trained STAN model.}
}
\value{
post is a list containing following posterior weights.
\item{beta} matrix containing source specific beta parameters.
\item{betaShared} vector of shared beta parameters common for all sources.
\item{tau} global scaling factor learned.
\item{sigma} noise parameter.
}
\description{
\code{posterior.bmsr.stan} extracts the posterior values from mode output.
}
