% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsr.R
\name{pred}
\alias{pred}
\title{predict function for single source}
\usage{
pred(xTest.S, beta, W, yN = NULL)
}
\arguments{
\item{xTest.S}{is a matrix of test data for predicting the outcome.}

\item{beta}{is model parameter vector.}

\item{W}{are model parameters. If NULL, model is single-task, else multi-task.}

\item{yN}{a list containing values used for normalizing the data: (default = NULL).
\itemize{
\item{cm}{vector of means with which the data is centered (0's if data is not centered),}
\item{cs}{vector of standard deviations with which the data is scaled (1's if data is not scaled).}
}}
}
\value{
yPred prediction vector of the stan model.
}
\description{
predict function for single source
}
\keyword{internal}
