% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsr.R
\name{predict.bmsmtr.stan}
\alias{predict.bmsmtr.stan}
\title{predict function for bmsmtr model}
\usage{
\method{predict}{bmsmtr.stan}(out, xTest, nTest, yN = NULL)
}
\arguments{
\item{out}{is trained STAN model.}

\item{xTest}{is a matrix of test data for predicting the outcome. If NULL no prediction is made (default).}

\item{nTest}{is a vector of length S, containing the number of values in each source. Can contain zero's.}

\item{yN}{a list containing values used for normalizing the data: (default = NULL).
\itemize{
\item{cm}{vector of means with which the data is centered, (0's if data is not centered),}
\item{cs}{vector of standard deviations with which the data is scaled (1's if data is not scaled).}
}}
}
\value{
yPred prediction vector of the bmsmtr model.
}
\description{
\code{predict.bmsmtr.stan} predicts the output of bmsmtr model, used as \code{predFunction} in \code{predictSTAN}.
}
