% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpingFunctions.R
\name{varNorm}
\alias{varNorm}
\title{variance normalize a data matrix}
\usage{
varNorm(mat, FFSC)
}
\arguments{
\item{mat}{matrix to variance normalize normalize.}

\item{FFSC}{the index of first column to standardize. All columns including this and after are normalized.}

\item{verbose}{print logs or not. Defaults to FALSE.}
}
\value{
a list containing the following:
\item{mat}{column-wise variance normalized matrix,}
\item{cs}{vector of column standard deviations used for variance normalization.}
}
\description{
\code{varNorm} performs column-wise standard-deviation=1 normalization of a matrix for all columns starting from \code{FFSC}
}
\keyword{internal}
