% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_bms.R
\name{VB_bms}
\alias{VB_bms}
\title{Bayesian model selection for group studies}
\usage{
VB_bms(m, n_samples = 1e+06)
}
\arguments{
\item{m}{N by K (N subjects by K models) matrix of log-model evidence.}

\item{n_samples}{Number of samples used in the simulation used for calculating exceedance probabilities.}
}
\value{
a list with the posterior estimates of the Dirichlet parameters (alpha), the expected model frequencies (r), the exceedance probabilities (xp), the Bayesian Omnibus Risk (bor), and the protected exceedance probabilities (pxp).
}
\description{
This function is translated from spm_BMS.m function, of SPM12, https://github.com/neurodebian/spm12/blob/master/spm_BMS.m
}
