% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpingFunctions.R
\name{ztransform}
\alias{ztransform}
\title{z-transform a training data matrix}
\usage{
ztransform(mat, FFSC, verbose = FALSE)
}
\arguments{
\item{mat}{matrix to z-transform}

\item{FFSC}{the index of first column to standardize. All columns including this and after are normalized.}

\item{verbose}{print logs or not. Defaults to FALSE.}
}
\value{
a list containing the following:
\item{mat} column-wise z-transformed matrix.
\item{cm} vector of column means used for z-transform.
\item{cs} vector of column standard deviations used for z-transform.
}
\description{
\code{ztransform} performs column-wise mean=0 and sd=1 normalization of a matrix for all columns starting from \code{FFSC}
}
